fix: added breakpoints to theme to allow for responsiveness on smaller mobile screens(iphone SE/6/7), updated newAuthPage to use grid
