+++ src/pages/authPage/SignUpForm.tsx
@@ -1,6 +1,7 @@
-import { Box, Button, TextField } from "@mui/material";
+import { Box, Button, TextField, useTheme } from "@mui/material";
 import React, { useState } from "react";
 import { userService } from "../../services/user";
+
 
 interface SignUpProps {
   setTab: React.Dispatch<React.SetStateAction<number>>;
@@ -15,6 +16,8 @@ interface FormData {
 
 const SignUpForm = (props: SignUpProps) => {
   const { setTab } = props;
+  const theme = useTheme();
+
 
   const [formData, setFormData] = useState<FormData>({
     userName: "",
@@ -120,6 +123,9 @@ const SignUpForm = (props: SignUpProps) => {
           fontWeight: "bold",
           px: 5,
           py: 1,
+          [theme.breakpoints.down("iphone7")]:{
+            mt:6
+          }
         }}
         color={"secondary"}
         variant={"contained"}
