+++ src/theme/theme.tsx
@@ -5,16 +5,27 @@ interface ThemeColors {
   primaryBackground: string;
   primaryText: string;
   primaryColor: string;
-  secondaryColor: string
-  contrastTextColor: string,
+  secondaryColor: string;
+  contrastTextColor: string;
   paper: string;
-  blue:string;
+  blue: string;
   // Add other common theme colors here
 }
 
 interface ThemeMode {
   light: ThemeColors;
   dark: ThemeColors;
+}
+
+declare module "@mui/material/styles" {
+  interface BreakpointOverrides {
+    xs: true; // removes the `xs` breakpoint
+    sm: true;
+    md: true;
+    lg: true;
+    xl: true;
+    iphone7: true; // adds the `mobile` breakpoint
+  }
 }
 
 const commonThemeColors: ThemeMode = {
@@ -25,7 +36,7 @@ const commonThemeColors: ThemeMode = {
     secondaryColor: "#000000",
     contrastTextColor: "#ffffff",
     paper: "white",
-    blue: "#000000"
+    blue: "#000000",
 
     // Define other common light theme colors here
   },
@@ -36,7 +47,7 @@ const commonThemeColors: ThemeMode = {
     secondaryColor: "#ffffff",
     contrastTextColor: "#000000",
     paper: "black",
-    blue: "#000000"
+    blue: "#000000",
 
     // Define other common dark theme colors here
   },
@@ -61,15 +72,25 @@ export const createAppTheme = (
         // Custom primary color for buttons and other components
         main: colors[themeMode].primaryColor,
         light: colors[themeMode].blue,
-        contrastText: colors[themeMode].contrastTextColor
+        contrastText: colors[themeMode].contrastTextColor,
       },
       secondary: {
         main: colors[themeMode].secondaryColor,
-        contrastText: colors[themeMode].contrastTextColor
+        contrastText: colors[themeMode].contrastTextColor,
+      },
 
-      }
       // You can customize other aspects of the theme here
       // typography, spacing, breakpoints, etc.
+    },
+    breakpoints: {
+      values: {
+        xs: 0,
+        sm: 600,
+        md: 900,
+        lg: 1200,
+        xl: 1536,
+        iphone7: 380,
+      },
     },
   });
 };
